+++ MyMoney.Web/ClientApp/src/app/app.module.ts
@@ -13,7 +13,7 @@ import {
    AddTransactionComponent,
    BudgetsComponent,
    AddBudgetComponent,
-   EditTransactionComponent,
+   EditBasicTransactionComponent,
    EditBudgetComponent,
    IncomesComponent,
    EditIncomeComponent,
@@ -51,7 +51,7 @@ import { routes } from './app.routing';
       LoginComponent,
       RegisterComponent,
       AddTransactionComponent,
-      EditTransactionComponent,
+      EditBasicTransactionComponent,
       TransactionsComponent,
       BudgetsComponent,
       AddBudgetComponent,
