+++ MyMoney.Web/ClientApp/src/app/pages/index.ts
@@ -1,21 +1,31 @@
 export * from './home/home.component';
+
+// User
+export * from './user/profile/profile.component';
+export * from './user/change-password/change-password.component';
+
+// Auth
+export * from './auth/forgot-password/forgot-password.component';
+export * from './auth/reset-password/reset-password.component';
 export * from './auth/login/login.component';
 export * from './auth/register/register.component';
-export * from './transactions/add-transaction/add-transaction.component';
-export * from './transactions/transactions.component';
+
+// Budget
 export * from './budgets/budgets.component';
 export * from './budgets/add-budget/add-budget.component';
-export * from './transactions/edit-transaction/edit-transaction.component';
 export * from './budgets/edit-budget/edit-budget.component';
+
+// Income
 export * from './incomes/incomes.component';
 export * from './incomes/add-income/add-income.component';
 export * from './incomes/edit-income/edit-income.component';
-export * from './transactions/import/import.transactions.component';
 export * from './incomes/import-incomes/import-incomes.component';
-export * from './user/profile/profile.component';
-export * from './auth/forgot-password/forgot-password.component';
-export * from './auth/reset-password/reset-password.component';
-export * from './user/change-password/change-password.component';
-export * from './transactions/edit-recurring-transaction/edit-recurring-transaction.component';
+
+// Transaction
+export * from './transactions/transactions.component';
+export * from './transactions/add-transaction/add-transaction.component';
 export * from './transactions/add-transaction/add-basic-transaction/add-basic-transaction.component';
 export * from './transactions/add-transaction/add-recurring-transaction/add-recurring-transaction.component';
+export * from './transactions/edit-transaction/edit-basic-transaction/edit-transaction.component';
+export * from './transactions/edit-transaction/edit-recurring-transaction/edit-recurring-transaction.component';
+export * from './transactions/import/import.transactions.component';
