+++ MyMoney.Web/ClientApp/src/app/pages/transactions/edit-transaction/edit-basic-transaction/edit-basic-transaction.component.html
