+++ MyMoney.Web/ClientApp/src/app/pages/transactions/edit-transaction/edit-recurring-transaction/edit-recurring-transaction.component.scss
